feat: add expected_cache_duplicate to optimize audio fetching
I am looking to the AudioGenerationTask datamodel in audio_generation/models to facilitate more intelligent fetching. We'll add a field "expected_cache_duplicate" which will allow us to skip fetching these clips in audio_generation/processing.py

In plan_audio_generation  in audio_generation/processing.py, we'll keep a list of cache filepaths. If the cache filepath is not in the list, expected_cache_duplicate will be false. If it is already in the list, expected_cache_duplicate will be true. 

In fetch_and_cache_audio in  audio_generation/processing.py, in the main loop, we'll continue the loop with a debug log note that we are skipping fetching due to the expectation that another fetch fill cache this path.

The rest of the logic should be unaffected. 

Let's also remember to update the appropriate tests in tests/audio_generation

You can read and write files using codemcp

codemcp-id: 15-feat-add-expected-cache-duplicate-to-optimize-audi