test: Update TTS provider tests to match refactored signatures
I have made a number of updates to method signatures, and a bit of logic, for the tts_providers module.

I have attached the list that includes the changes that have been made. Can you update the tests in the tests/tts_providers module to match the new method signatures / logic? There are possible places where the actual methods need to be updated, but we should look to fixing the tests first.

codemcp-id: 17-test-update-tts-provider-tests-to-match-refactored